requirements.txt
google-cloud-bigquery>=3.0.0
pandas>=2.2.3
protobuf>=6.31.0
requests>=2.31.0
python-dotenv>=1.0.0